backend/requirements.txt
asgiref==3.6.0
cffi==1.15.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
zope.event==4.6
zope.interface==5.5.2
